fix: version announcement modal
